function: implement add_init_scripts for install/base
When we are building mkinitcpio while developing we are never really
using the in-development init and functions from the source, but
including out global files from the actual install.

This checks if we are sourcing the function file by name or not to
figure out which files we are including.

Signed-off-by: Morten Linderud <morten@linderud.pw>