libalpm/scripts/mkinitcpio: rebuild initramfs images when packages needed by a hook change
Create an associative array containing mkinitcpio hook names and file
paths of files used by hooks. When a file in one of the paths changes as
part of a package install/upgrade/etc., the script will check all
mkinitcpio presets and rebuild the initramfs for all that use hooks which
rely on these changed files.