Revert "mkinitcpio: fix behavior generatedir"
This reverts commit 01c71e23743ceab44e1fd5bd8cbbe01b9f1aaa77.

This commit changed the semantics of initialize_buildroot and made some
invocations of the function leave garbage in /tmp/mkinitcpio.XXXXXX,
especially during tests. Restore the previous behaviour, and instead
copy the built tree to _opttargetdir if the flag is passed. This should
keep things working for mkinitcpio-generate-shutdown-ramfs.service.

Alternate working directories are always passed with -t, which exports a
different TMPDIR, so there should be no need for the second argument on
initialize_buildroot.